Reorganize: give the two drivers their shared vocabulary in one file
activityConfig lived in the standalone driver and activityInfo in the
workflow one, though neither belongs to a surface: the workflow driver
reached across for eight shared things and the standalone one reached back
for activityInfo. Which file held which was arbitrary, and each driver
could only be read with the other open.

activity_driver.go now holds what they share, so each surface file holds
only that surface.

  go test -count=1 -tags test_dep -run TestActivityParityTestSuite ./tests/

Pass the activity config to the helper workflow

wfaActivityParams restated nine of activityConfig's fields, renaming one,
so adding an option meant touching the struct, the mapping in start, and
the ActivityOptions in the workflow. It now carries the config itself plus
where to schedule the activity.

That also puts the defaults in one place. start resolved start-to-close and
the retry interval while the workflow separately re-derived the coefficient
and the maximum interval; all four now come from the same accessors
startRequest uses, so the two surfaces cannot drift.

  go test -count=1 -tags test_dep -run TestActivityParityTestSuite ./tests/
